+++ Common/src/main/kotlin/gay/object/hexdebug/casting/eval/SplicingTableCastEnv.kt
@@ -7,6 +7,7 @@ import at.petrak.hexcasting.api.casting.eval.vm.CastingImage
 import at.petrak.hexcasting.api.pigment.FrozenPigment
 import at.petrak.hexcasting.api.utils.extractMedia
 import at.petrak.hexcasting.common.lib.hex.HexEvalSounds
+import gay.`object`.hexdebug.blocks.splicing.SplicingTableBlock
 import gay.`object`.hexdebug.blocks.splicing.SplicingTableBlockEntity
 import gay.`object`.hexdebug.config.HexDebugServerConfig
 import net.minecraft.core.BlockPos
@@ -14,7 +15,6 @@ import net.minecraft.core.Direction
 import net.minecraft.server.level.ServerPlayer
 import net.minecraft.sounds.SoundSource
 import net.minecraft.world.InteractionHand
-import net.minecraft.world.level.block.state.properties.BlockStateProperties
 import net.minecraft.world.phys.Vec3
 import kotlin.math.min
 
@@ -26,7 +26,7 @@ class SplicingTableCastEnv(
 
     val blockPos: BlockPos get() = table.blockPos
 
-    val facing: Direction get() = table.blockState.getValue(BlockStateProperties.FACING)
+    val facing: Direction get() = table.blockState.getValue(SplicingTableBlock.FACING)
 
     override fun postExecution(result: CastResult) {
         super.postExecution(result)
@@ -35,7 +35,6 @@ class SplicingTableCastEnv(
 
     override fun postCast(image: CastingImage) {
         super.postCast(image)
-        table.sync() // TODO: is this necessary?
         sound.sound?.let {
             world.playSound(null, blockPos, it, SoundSource.PLAYERS, 1f, 1f)
         }
@@ -57,13 +56,9 @@ class SplicingTableCastEnv(
 
         // then, pull straight from the item in the media slot
         if (costLeft > 0) {
-            HexAPI.instance().findMediaHolder(table.mediaStack)?.let {
+            table.mediaHolder?.let {
                 costLeft -= extractMedia(it, cost = costLeft, simulate = simulate)
             }
-        }
-
-        if (!simulate) {
-            table.refillMedia()
         }
 
         return costLeft