+++ Common/src/main/kotlin/gay/object/hexdebug/config/HexDebugClientConfig.kt
@@ -4,6 +4,7 @@ import at.petrak.hexcasting.api.utils.asTranslatedComponent
 import com.mojang.blaze3d.platform.InputConstants
 import gay.`object`.hexdebug.HexDebug
 import gay.`object`.hexdebug.gui.splicing.SplicingTableScreen
+import gay.`object`.hexdebug.splicing.SplicingTableAction
 import me.shedaniel.autoconfig.AutoConfig
 import me.shedaniel.autoconfig.ConfigData
 import me.shedaniel.autoconfig.ConfigHolder
@@ -37,12 +38,12 @@ object HexDebugClientConfig {
                     entryBuilder.startModifierKeyCodeField(
                         when (config) {
                             is ClientConfig.SplicingTableKeybinds,
-                            is ClientConfig.EnlightenedSplicingTableKeybinds -> {
-                                SplicingTableScreen.buttonText(field.name.camelToSnakeCase())
+                            is ClientConfig.SplicingTableKeybinds.Enlightened -> {
+                                SplicingTableScreen.buttonText(field.name.camelToSnakeCase(), null)
                             }
                             else -> i18n.asTranslatedComponent
                         },
-                        getUnsafely(field, config, ConfigModifierKey(InputConstants.UNKNOWN.name)).inner,
+                        getUnsafely(field, config, ConfigModifierKey()).inner,
                     )
                         .setModifierDefaultValue { (getUnsafely(field, defaults) as ConfigModifierKey).inner }
                         .setModifierSaveConsumer { setUnsafely(field, config, ConfigModifierKey(it)) }
@@ -99,27 +100,95 @@ object HexDebugClientConfig {
         @CollapsibleObject
         val splicingTableKeybinds = SplicingTableKeybinds()
 
-        @Tooltip
-        @CollapsibleObject
-        val enlightenedSplicingTableKeybinds = EnlightenedSplicingTableKeybinds()
-
+        @Suppress("MemberVisibilityCanBePrivate")
         class SplicingTableKeybinds {
+            @Tooltip
+            val enabled: Boolean = true
+
+            @Tooltip
+            val overrideVanillaArrowKeys = true
+
+            val viewLeft = ConfigModifierKey(InputConstants.KEY_UP)
+            val viewLeftPage = ConfigModifierKey(InputConstants.KEY_PAGEUP)
+            val viewLeftFull = ConfigModifierKey(InputConstants.KEY_HOME)
+            val viewRight = ConfigModifierKey(InputConstants.KEY_DOWN)
+            val viewRightPage = ConfigModifierKey(InputConstants.KEY_PAGEDOWN)
+            val viewRightFull = ConfigModifierKey(InputConstants.KEY_END)
+
+            val cursorLeft = ConfigModifierKey(InputConstants.KEY_LEFT)
+            val cursorRight = ConfigModifierKey(InputConstants.KEY_RIGHT)
+
+            val expandSelectionLeft = ConfigModifierKey(InputConstants.KEY_LEFT, shift = true)
+            val expandSelectionRight = ConfigModifierKey(InputConstants.KEY_RIGHT, shift = true)
+            val moveSelectionLeft = ConfigModifierKey(InputConstants.KEY_LEFT, ctrl = true, shift = true)
+            val moveSelectionRight = ConfigModifierKey(InputConstants.KEY_RIGHT, ctrl = true, shift = true)
+
             val selectNone = ConfigModifierKey(InputConstants.KEY_A, ctrl = true, shift = true)
             val selectAll = ConfigModifierKey(InputConstants.KEY_A, ctrl = true)
+
             val undo = ConfigModifierKey(InputConstants.KEY_Z, ctrl = true)
             val redo = ConfigModifierKey(InputConstants.KEY_Y, ctrl = true)
+
             val nudgeLeft = ConfigModifierKey(InputConstants.KEY_LEFT, ctrl = true)
             val nudgeRight = ConfigModifierKey(InputConstants.KEY_RIGHT, ctrl = true)
+
             val duplicate = ConfigModifierKey(InputConstants.KEY_D, ctrl = true)
             val delete = ConfigModifierKey(InputConstants.KEY_DELETE)
+            val backspace = ConfigModifierKey(InputConstants.KEY_BACKSPACE)
+
             val cut = ConfigModifierKey(InputConstants.KEY_X, ctrl = true)
             val copy = ConfigModifierKey(InputConstants.KEY_C, ctrl = true)
             val pasteSplat = ConfigModifierKey(InputConstants.KEY_V, ctrl = true)
             val pasteVerbatim = ConfigModifierKey(InputConstants.KEY_V, ctrl = true, shift = true)
-        }
 
-        class EnlightenedSplicingTableKeybinds {
-            val cast = ConfigModifierKey(InputConstants.KEY_RETURN, ctrl = true)
+            @Tooltip
+            @CollapsibleObject(startExpanded = true)
+            val enlightened = Enlightened()
+
+            fun getActionMap() = mapOf(
+                SplicingTableAction.VIEW_LEFT to viewLeft,
+                SplicingTableAction.VIEW_LEFT_PAGE to viewLeftPage,
+                SplicingTableAction.VIEW_LEFT_FULL to viewLeftFull,
+                SplicingTableAction.VIEW_RIGHT to viewRight,
+                SplicingTableAction.VIEW_RIGHT_PAGE to viewRightPage,
+                SplicingTableAction.VIEW_RIGHT_FULL to viewRightFull,
+                SplicingTableAction.CURSOR_LEFT to cursorLeft,
+                SplicingTableAction.CURSOR_RIGHT to cursorRight,
+                SplicingTableAction.EXPAND_SELECTION_LEFT to expandSelectionLeft,
+                SplicingTableAction.EXPAND_SELECTION_RIGHT to expandSelectionRight,
+                SplicingTableAction.MOVE_SELECTION_LEFT to moveSelectionLeft,
+                SplicingTableAction.MOVE_SELECTION_RIGHT to moveSelectionRight,
+                SplicingTableAction.SELECT_NONE to selectNone,
+                SplicingTableAction.SELECT_ALL to selectAll,
+                SplicingTableAction.UNDO to undo,
+                SplicingTableAction.REDO to redo,
+                SplicingTableAction.NUDGE_LEFT to nudgeLeft,
+                SplicingTableAction.NUDGE_RIGHT to nudgeRight,
+                SplicingTableAction.DUPLICATE to duplicate,
+                SplicingTableAction.DELETE to delete,
+                SplicingTableAction.BACKSPACE to backspace,
+                SplicingTableAction.CUT to cut,
+                SplicingTableAction.COPY to copy,
+                SplicingTableAction.PASTE_SPLAT to pasteSplat,
+                SplicingTableAction.PASTE_VERBATIM to pasteVerbatim,
+            )
+
+            fun getActionForKey(keyCode: Int, scanCode: Int): SplicingTableAction? {
+                for ((action, key) in getActionMap().entries) {
+                    if (key.inner.matchesKey(keyCode, scanCode)) {
+                        return action
+                    }
+                }
+                return null
+            }
+
+            fun getKeyForAction(action: SplicingTableAction): ConfigModifierKey? {
+                return getActionMap()[action]
+            }
+
+            class Enlightened {
+                val cast = ConfigModifierKey(InputConstants.KEY_RETURN, ctrl = true)
+            }
         }
     }
 }
@@ -165,8 +234,6 @@ data class ConfigModifierKey(
             Modifier.of(alt, ctrl, shift),
         )
     }
-
-    fun matchesKey(keyCode: Int, scanCode: Int) = inner.matchesKey(keyCode, scanCode)
 }
 
 // https://stackoverflow.com/a/60010299